Added hasher to the type of the s_proxyTypeToName variable. (#6)
* Added hasher to the type of the s_proxyTypeToName unordered_map variable.

This fixes compile issue that compiler is unable to create hash for the key that is
enum in this case.

* Replaced CXX_STANDARD property with cxx_std_11 feature.

Setting of CXX_STANDARD does not allow to change the standard via CMake command line with
DCMAKE_CXX_STANDARD command line parameter. On the other hand cxx_std_11 forces minimal C++
standard version.